+++ src/main/java/uni/gaben/iscat/gamenex/universe/iscat_mob/IscatMobSettings.java
@@ -21,7 +21,7 @@ public class IscatMobSettings {
     /** Punti vita iniziali del mob. */
     public static final int HP_INIZIALI = 30;
     /** Dimensione base dello sprite in pixel. */
-    public static final double DIM_SPRITE = 32.0;
+    public static final int DIM_SPRITE = 32;
     /** Numero di frame dell'animazione. */
     public static final int NUMERO_FRAMES = 1;
     /** Scala di rendering. */
